Commit 5: Created typed clients
